% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cck.R
\name{cck}
\alias{cck}
\title{Applys CCK to the correlations for a function g}
\usage{
cck(dat, g, translate = cor_vec, alpha = 0.05, trials = 100, cores = 2)
}
\arguments{
\item{dat}{nxd numeric matrix}

\item{g}{function that takes in (d choose 2) numbers}

\item{translate}{translates each sample of length d into some other vector}

\item{alpha}{significance value, between 0 and 1}

\item{trials}{positive integer of bootstrap trials}

\item{cores}{number of cores to parallelize computation over}
}
\description{
Applys CCK to the correlations for a function g
}
