% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stepdown.R
\name{connected_components}
\alias{connected_components}
\title{Returns the number of connected components from a list of edges}
\usage{
connected_components(d, edges)
}
\arguments{
\item{d}{number of nodes}

\item{edges}{2x(number of edges) matrix of integers}
}
\value{
list of the connected components
}
\description{
Returns the number of connected components from a list of edges
}
