% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cck.R
\name{cor_vec}
\alias{cor_vec}
\title{Converts the data into correlations}
\usage{
cor_vec(dat, sigma_vec = rep(1, ncol(dat)), noise_vec = rep(1, nrow(dat)))
}
\arguments{
\item{dat}{nxd numeric matrix}

\item{sigma_vec}{length d numeric vector}

\item{noise_vec}{length d numeric vector}
}
\value{
a vector of length (d choose 2).
}
\description{
Allows \code{noise_vec} which multiplies each sample with a certain amount of
noise. Also requires the estimated standard deviations (marginal) in
\code{sigma_vec}.
}
