% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gLatentModel.R
\name{gLatentModel}
\alias{gLatentModel}
\title{Estimate G Latent Model}
\usage{
gLatentModel(dat, K, ...)
}
\arguments{
\item{dat}{n by d matrix where there are n samples and d variables}

\item{K}{number of clusters for the d variables}

\item{...}{additional parameters for binary search or cord}
}
\value{
a gLatentModel object
}
\description{
Estimate G Latent Model
}
