% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gLatentModelSBM.R
\name{gLatentModel_sbm}
\alias{gLatentModel_sbm}
\title{Estimate G Latent Model}
\usage{
gLatentModel_sbm(dat, K, debugging = F, ...)
}
\arguments{
\item{dat}{n by d matrix where there are n samples and d variables}

\item{K}{number of clusters for the d variables}

\item{debugging}{boolean on whether or not to pass additional parameters out}

\item{...}{additional parameters for binary search or cord}
}
\value{
a gLatentModel object
}
\description{
Estimate G Latent Model
}
