% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/naiveMethods.R
\name{naive_clustering_hclust}
\alias{naive_clustering_hclust}
\title{Variable clustering via Hierarchical Clustering}
\usage{
naive_clustering_hclust(dat, K)
}
\arguments{
\item{dat}{data matrix n x d}

\item{K}{the number of clusters}
}
\value{
a list containing the latent covariance matrix and the clustering
}
\description{
Variable clustering via Hierarchical Clustering
}
