% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/naiveMethods.R
\name{naive_clustering_sbm}
\alias{naive_clustering_sbm}
\title{Variable clustering via Stochastic Block Model}
\usage{
naive_clustering_sbm(dat, K)
}
\arguments{
\item{dat}{data matrix n x d}

\item{K}{the number of clusters}
}
\value{
a list containing the latent covariance matrix and the clustering
}
\description{
Motivated from https://projecteuclid.org/euclid.aos/1418135620
}
