% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance.R
\name{partition_distance}
\alias{partition_distance}
\title{Computes the distance between two partitions.}
\usage{
partition_distance(partition1, partition2, method = "variation",
  covariance = NA)
}
\arguments{
\item{partition1}{a partition vector}

\item{partition2}{a partition vector}

\item{method}{a string}

\item{covariance}{a PSD matrix}
}
\value{
a numeric
}
\description{
\code{partition1} and \code{partition2} must be valid partitions. See
\code{.check_partition} to see what this means.
You can also pass in a covariance matrix (to take the distance with
respect to), but this is only required for certain \code{method}s.
}
