% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cck.R
\name{row_difference_closure}
\alias{row_difference_closure}
\title{Generate function to evaluate difference between two rows}
\usage{
row_difference_closure(i, j, d)
}
\arguments{
\item{i}{integer between 1 and d}

\item{j}{integer between 1 and d}

\item{d}{positive ingeter}
}
\value{
a function
}
\description{
Generate function to evaluate difference between two rows
}
