% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stepdown.R
\name{stepdown}
\alias{stepdown}
\title{Stepdown procedure}
\usage{
stepdown(dat, g_list, translate = cor_vec, alpha = 0.05, trials = 100,
  cores = 2)
}
\arguments{
\item{dat}{nxd numeric matrix}

\item{g_list}{a list of function that takes in (d choose 2) numbers}

\item{translate}{translates each sample of length d into some other vector}

\item{alpha}{significance value, between 0 and 1}

\item{trials}{positive integer of bootstrap trials}

\item{cores}{number of cores to parallelize computation over}
}
\value{
a set of indices of integers between 1 to \code{length(g_list)}
to denote all the hypothesis that passed
}
\description{
Stepdown procedure
}
