% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gLatentModel.R
\name{gLatentModel}
\alias{gLatentModel}
\title{Estimate G Latent Model}
\usage{
gLatentModel(dat, K, verbose = F, seed = NA, num_subsample = NA)
}
\arguments{
\item{dat}{n by d matrix where there are n samples and d variables}

\item{K}{number of clusters for the d variables}

\item{verbose}{boolean to print out progress statements}

\item{seed}{seed to fix the random subsampling and kmeans within estimation}

\item{num_subsample}{number of subsamples when initially estimate variance of noise}
}
\value{
a gLatentModel object
}
\description{
Estimate G Latent Model
}

